+++ geolocator_apple/ios/Classes/Constants/ErrorCodes.m
@@ -11,6 +11,7 @@ NSString * const GeolocatorErrorLocationUpdateFailure = @"LOCATION_UPDATE_FAILUR
 NSString * const GeolocatorErrorLocationServicesDisabled = @"LOCATION_SERVICES_DISABLED";
 NSString * const GeolocatorErrorLocationSubscriptionActive = @"LOCATION_SUBSCRIPTION_ACTIVE";
 NSString * const GeolocatorErrorLocationSignalLost = @"LOCATION_SIGNAL_LOST";
+NSString * const GeolocatorErrorLocationSubscriptionInactive = @"LOCATION_SUBSCRIPTION_INACTIVE";
 NSString * const GeolocatorErrorPermissionDefinitionsNotFound = @"PERMISSION_DEFINITIONS_NOT_FOUND";
 NSString * const GeolocatorErrorPermissionDenied = @"PERMISSION_DENIED";
 NSString * const GeolocatorErrorPermissionRequestInProgress = @"PERMISSION_REQUEST_IN_PROGRESS";