+++ geolocator_apple/lib/src/geolocator_apple.dart
@@ -199,6 +199,22 @@ class GeolocatorApple extends GeolocatorPlatform {
   }
 
   @override
+  Future<void> updatePositionStream(
+      {required LocationSettings locationSettings}) async {
+    if (_positionStream == null) {
+      throw const NotSubscribedException();
+    }
+    try {
+      await _methodChannel.invokeMethod(
+          'updatePositionStream', locationSettings.toJson());
+    } on PlatformException catch (e) {
+      final error = _handlePlatformException(e);
+
+      throw error;
+    }
+  }
+
+  @override
   Future<LocationAccuracyStatus> requestTemporaryFullAccuracy({
     required String purposeKey,
   }) async {
@@ -236,6 +252,8 @@ class GeolocatorApple extends GeolocatorPlatform {
         return const AlreadySubscribedException();
       case 'LOCATION_SIGNAL_LOST':
         return const LocationSignalLostException();
+      case 'LOCATION_SUBSCRIPTION_INACTIVE':
+        return const NotSubscribedException();
       case 'PERMISSION_DEFINITIONS_NOT_FOUND':
         return PermissionDefinitionsNotFoundException(exception.message);
       case 'PERMISSION_DENIED':